Separação das classes em arquivos cpp e header
Fiz a separação das classes em arquivos cpp e header para fins de organização
